+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/routealternatives/RouteAlternativesController.kt
@@ -240,7 +240,7 @@ internal class RouteAlternativesController constructor(
     }
 }
 
-private fun RouteAlternative.mapToMetadata(
+internal fun RouteAlternative.mapToMetadata(
     navigationRoute: NavigationRoute
 ): AlternativeRouteMetadata {
     return AlternativeRouteMetadata(